fix: detached args at the end of chain access
Note that we could not put initials on the first line after this.
